React FINALLY rendering properly
