+++ client/src/App.js
@@ -6,21 +6,50 @@ import Home from './pages/Home.js';
 import Login from './pages/Login';
 import NavBar from './components/NavBar';
 import Footer from './components/Footer';
+import { setContext } from '@apollo/client/link/context';
+import {
+  ApolloClient,
+  InMemoryCache,
+  ApolloProvider,
+  createHttpLink,
+} from "@apollo/client";
 
+
+const httpLink = createHttpLink({
+  uri: "/graphql",
+});
+
+const authLink = setContext((_, { headers }) => {
+  const token = localStorage.getItem("id_token");
+  return {
+    headers: {
+      ...headers,
+      authorization: token ? `Bearer ${token}` : "",
+    },
+  };
+});
+
+const client = new ApolloClient({
+  link: authLink.concat(httpLink),
+  cache: new InMemoryCache(),
+});
 
 function App() {
   return (
-  <div>
+    <ApolloProvider client={client}>
+  <div> 
     <NavBar />
    
     <Routes>
-      <Route index element={<Home />} />
-      <Route path="MyBar" element={<MyBar />}/>
-      <Route path="Login" element={<Login />}/>
+      <Route exact path="/" element={<Home />} />
+      <Route exact path="/mybar" element={<MyBar />}/>
+      <Route exact path="/login" element={<Login />}/>
     </Routes>
-    
+    <h1>Hello</h1>
     <Footer />
+    
     </div>
+    </ApolloProvider>
   )
 }
