+++ client/src/Pages/Home.js
@@ -3,7 +3,7 @@ import React from 'react';
 const Home = () => {
     return (
         <main>
-            <Home />
+            <h1>Home</h1>
         </main>
     )
 }
